requirements.txt
# AutoStream AI Agent — Python Dependencies
# Install with: pip install -r requirements.txt

# LLM Framework
langchain
langchain-core
langchain-openai
langgraph

# Utilities
python-dotenv     
pytest
